feat : map_insert works
There you will have to use --unwind 3 flag with the esbmc to make use of maps. otherwise it will run into a VERY long loop unrolling process.

this branch also introduces top_objtype variable that contains the topmost objtype in the node. Can be referenced anywhere in the code as and when needed.

It also introcues capability to add code to constructor.
- look for map_init_block and move_mapping_to_ctor