feat: adds var-get translation (#23)
* feat : map_init working

requires manual entry of mapping types from the python script output to clarity_template.h

* feat : map_insert works

There you will have to use --unwind 3 flag with the esbmc to make use of maps. otherwise it will run into a VERY long loop unrolling process.

this branch also introduces top_objtype variable that contains the topmost objtype in the node. Can be referenced anywhere in the code as and when needed.

It also introcues capability to add code to constructor.
- look for map_init_block and move_mapping_to_ctor

* Chore : minor fixes

* fix : converted map_insert code to a function

* fix : added missing return type

* feat: adds var-get translation

* adds template functions for primitives and sequences
* sets custom type properties for all literals

* feat: adds lists handling

* fixes string type inference (utf8 vs ascii)
* adds list sub struct type
* fixes list length
* renamed principal struct params

* fix: fixes utf8 handling and PR feedback

---------

Co-authored-by: Muhammad Ali <ali@tintash.com>